refactor: refactored and restructured notificationPreferences page
