+++ src/index.scss
@@ -95,6 +95,14 @@ $fa-font-path: "~font-awesome/fonts";
     height: 36px;
     font-weight: 500;
   }
+
+  .line-height-36{
+    line-height: 36px;
+  }
+
+  .h-4\.5 {
+    height: 36px;
+  }
 }
 
 .usabilla_live_button_container {
